requirements.txt
fastf1>=3.1.0
kafka-python>=2.0.2
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0
psutil>=5.9.0
flask>=3.0.0
flask-cors>=4.0.0
